refactor(header-menu): move files to features folder, remove commented code
